requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Click==8.0.0
Django==3.2.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-perms==2.1.0
django-recaptcha3==0.1.0
django-smtp-ssl==1.0
djangorestframework==3.12.4
h11==0.12.0
httptools==0.2.0
idna==3.1
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4